Rename package, test/verify generated categories (api,model,infra)
