+++ .github/workflows/sbt.yml
@@ -24,4 +24,4 @@ jobs:
         path: ~/.sbt
         key: ${{ runner.os }}-sbt-${{ hashFiles('**/build.sbt') }}
     - name: Run tests
-      run: sbt test
+      run: sbt 'test; scripted;'
