+++ src/main/scala/org/openapitools/generator/sbt/plugin/tasks/OpenApiGenerateTask.scala
@@ -1,9 +1,9 @@
 /*
- * Copyright 2018 OpenAPI-Generator Contributors (https://openapi-generator.tech)
- * Copyright 2018 SmartBear Software
+ * Copyright (c) 2020 OpenAPI-Generator Contributors (https://openapi-generator.tech)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
+ *
  * You may obtain a copy of the License at
  *
  *     https://www.apache.org/licenses/LICENSE-2.0
@@ -14,14 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.openapitools.generator.sbt.plugin.tasks
 
-package sbtopenapigenerator.tasks
-
-import org.openapitools.codegen.{CodegenConstants, DefaultGenerator}
 import org.openapitools.codegen.config.{CodegenConfigurator, GlobalSettings}
-import sbt.{Def, _}
+import org.openapitools.codegen.{CodegenConstants, DefaultGenerator}
+import org.openapitools.generator.sbt.plugin.OpenApiGeneratorKeys
 import sbt.Keys._
-import sbtopenapigenerator.OpenApiGeneratorKeys
+import sbt.{Def, _}
 
 import scala.collection.JavaConverters._
 import scala.util.{Failure, Success, Try}
@@ -260,7 +259,7 @@ trait OpenApiGenerateTask extends OpenApiGeneratorKeys {
           }
         case Failure(ex) =>
           logger.info(ex.getMessage)
-          Seq()
+          Seq.empty
       }
     }
   }
